backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langgraph>=0.0.20
langsmith>=0.0.80
chromadb>=0.4.22
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
numpy>=1.24.0
networkx>=3.1
neo4j>=5.14.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pillow>=10.0.0
